Firebaseの設定を分離
