+++ src/firebase.js
@@ -1,0 +1,12 @@
+import firebase from 'firebase/app';
+
+const firebaseConfig = {
+  apiKey: "AIzaSyCScbzQJOJwndGPyU3Db_wAsWtmV9i6cLM",
+  authDomain: "myochikirin-hp-14b0d.firebaseapp.com",
+  projectId: "myochikirin-hp-14b0d",
+  storageBucket: "myochikirin-hp-14b0d.appspot.com",
+  messagingSenderId: "1078840520799",
+  appId: "1:1078840520799:web:9b14c7a2efd355938babea"
+};
+// Initialize Firebase
+firebase.initializeApp(firebaseConfig);
