+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import firebase from 'firebase/app';
+import './firebase'
 import './assets/css/index.css';
 import './assets/css/fontello.css'
 import App from './App';
@@ -12,17 +12,6 @@ ReactDOM.render(
   </React.StrictMode>,
   document.getElementById('root')
 );
-
-const firebaseConfig = {
-  apiKey: "AIzaSyCScbzQJOJwndGPyU3Db_wAsWtmV9i6cLM",
-  authDomain: "myochikirin-hp-14b0d.firebaseapp.com",
-  projectId: "myochikirin-hp-14b0d",
-  storageBucket: "myochikirin-hp-14b0d.appspot.com",
-  messagingSenderId: "1078840520799",
-  appId: "1:1078840520799:web:9b14c7a2efd355938babea"
-};
-// Initialize Firebase
-firebase.initializeApp(firebaseConfig);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
